+++ CHANGES.md
@@ -4,8 +4,8 @@
 - deal with "lost" clients, that disconnected, but didn't get added to the 
   connection log (issue #70)
 - `update-ip` script now assigns a /64 instead of /60 to IPv6 profile address
-- introduce `last_seen_web` to keep track of when the user was last seen at
-  the portal, or authorized an application
+- introduce `last_authenticated_at` to keep track of when the user last 
+  authenticated at the portal
 - remove the ability to disable certificates, only users can be disabled now
 
 ## 1.2.14 (2018-06-13)